@@ -12,19 +12,27 @@
     "@expo-google-fonts/jolly-lodger": "^0.2.3",
     "@expo-google-fonts/open-sans": "^0.2.3",
     "@expo/vector-icons": "^14.0.4",
+    "@react-navigation/bottom-tabs": "^7.2.1",
     "@react-navigation/native-stack": "^7.2.0",
-    "expo": "~52.0.37",
-    "expo-font": "~13.0.4",
-    "expo-splash-screen": "~0.29.22",
-    "expo-status-bar": "~2.0.1",
-    "react": "18.3.1",
-    "react-native": "0.76.7",
-    "react-native-safe-area-context": "4.12.0",
-    "react-native-screens": "~4.4.0"
+    "axios": "^1.8.4",
+    "expo": "^53.0.0",
+    "expo-font": "~13.3.1",
+    "expo-secure-store": "~14.2.3",
+    "expo-splash-screen": "^0.30.8",
+    "expo-status-bar": "~2.2.3",
+    "jwt-decode": "^4.0.0",
+    "react": "^19.0.0",
+    "react-native": "^0.79.2",
+    "react-native-gesture-handler": "~2.24.0",
+    "react-native-reanimated": "~3.17.4",
+    "react-native-safe-area-context": "^5.4.0",
+    "react-native-screens": "~4.10.0",
+    "react-native-svg": "^15.11.2",
+    "react-native-toast-message": "^2.2.1"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
-    "@types/react": "~18.3.12",
+    "@types/react": "~19.0.10",
     "typescript": "^5.3.3"
   },
   "private": true
