@@ -1,25 +1,20 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useCallback, useEffect } from "react";
 import * as SplashScreen from "expo-splash-screen";
-import * as Font from "expo-font";
-import LoginScreen from "./components/pages/InitialUserPages/login";
-import SingUp from "./components/pages/InitialUserPages/sign-up";
+
 import {
   useFonts,
   OpenSans_400Regular as OpenSans,
   OpenSans_700Bold as OpenSansBold,
 } from "@expo-google-fonts/open-sans";
 import { JollyLodger_400Regular as JollyLodger } from "@expo-google-fonts/jolly-lodger";
-import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
 
 import { Keyboard, TouchableWithoutFeedback, View } from "react-native";
-import UserProfile from "./components/pages/UserProfile/user-profile";
-
+import Routes from "./routes";
+import { GestureHandlerRootView } from "react-native-gesture-handler";
 
 SplashScreen.preventAutoHideAsync();
-const Stack = createNativeStackNavigator();
 
-export default function App() {
+const App: React.FC = () => {
   const [fontsLoaded] = useFonts({
     OpenSans,
     OpenSansBold,
@@ -37,16 +32,14 @@ export default function App() {
   }
 
   return (
+    <GestureHandlerRootView style={{ flex: 1 }}>
     <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-    <View style={{ flex: 1 }} onLayout={onLayoutRootView}>
-      <NavigationContainer>
-        <Stack.Navigator screenOptions={{ headerShown: false }}>
-          <Stack.Screen name="UserProfile" component={UserProfile} />
-          {/* <Stack.Screen name="Login" component={LoginScreen} />
-          <Stack.Screen name="SingUp" component={SingUp}  /> */}
-        </Stack.Navigator>
-      </NavigationContainer>
-    </View>
+      <View style={{ flex: 1 }} onLayout={onLayoutRootView}>
+        <Routes />
+      </View>
     </TouchableWithoutFeedback>
+    </GestureHandlerRootView>
   );
 }
+
+export default App;
