@@ -1,12 +1,71 @@
-import React from "react";
-import { View, TouchableOpacity, Text, TextInput } from "react-native";
+import React, { useEffect, useState } from "react";
+import {
+  View,
+  TouchableOpacity,
+  Text,
+  TextInput,
+  ActivityIndicator,
+} from "react-native";
 
 import styles from "./styles";
 import { ParamListBase, useNavigation } from "@react-navigation/native";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
+import Input from "../../atoms/input";
+import FormButton from "../../atoms/button";
+import {
+  getUserLoged,
+  login,
+  removeUserLoged,
+  storeUserLoged,
+} from "../../../services/user.service";
+import { LoginDTO } from "../../../models/login-data.model";
+import { isTokenExpired } from "../../../services/token-service";
 
 const LoginScreen = () => {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [loading, setLoading] = useState(false);
   const navigation = useNavigation<NativeStackNavigationProp<ParamListBase>>();
+
+  useEffect(() => {
+    const verifyUser = async () => {
+      console.log("Verificando usuário...");
+      if (await isTokenExpired()) {
+        await removeUserLoged();
+      }
+      
+      const user = await getUserLoged()
+
+      if (user) {
+        navigation.reset({
+          index: 0,
+          routes: [{ name: 'Tabs' }],
+        });
+      }
+    };
+
+    verifyUser();
+  }, []);
+
+  async function handleLogin() {
+    const loginDTO: LoginDTO = { email, password };
+
+    setLoading(true);
+    try {
+      const { data } = await login(loginDTO);
+
+      await storeUserLoged(data);
+      navigation.navigate("Tabs");
+    } catch (error: any) {
+      console.log(error);
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  if (loading) {
+    return <ActivityIndicator size="large" color="#0000ff" />;
+  }
 
   return (
     <View style={styles.container}>
@@ -19,28 +78,21 @@ const LoginScreen = () => {
       </View>
 
       <View style={styles.form}>
-        <TextInput
-          style={styles.input}
-          placeholder="E-mail"
-          placeholderTextColor="#888"
-        />
-        <TextInput
-          style={styles.input}
+        <Input placeholder="E-mail" value={email} onChangeText={setEmail} />
+        <Input
           placeholder="Senha"
-          placeholderTextColor="#888"
           secureTextEntry
+          value={password}
+          onChangeText={setPassword}
         />
-        <TouchableOpacity style={styles.button}>
-          <Text style={styles.buttonText}>Entrar</Text>
-        </TouchableOpacity>
+        <FormButton title="Entrar" onPress={handleLogin} />
       </View>
 
-      <TouchableOpacity
-        style={styles.registerButton}
+      <FormButton
+        title="Cadastrar-se"
         onPress={() => navigation.navigate("SingUp")}
-      >
-        <Text style={styles.registerText}>Cadastrar-se</Text>
-      </TouchableOpacity>
+        isOutline
+      />
     </View>
   );
 };
